Conflicts:
	package.json
	red/api/theme.js

+++ red/storage/localfilesystem.js
@@ -158,7 +158,9 @@ var localfilesystem = {
                 settings.userDir = process.env.NODE_RED_HOME;
             } else {
                 settings.userDir = fspath.join(process.env.HOME || process.env.HOMEPATH || process.env.USERPROFILE || process.env.NODE_RED_HOME,".node-red");
-                promises.push(promiseDir(settings.userDir));
+                if (!settings.readOnly) {
+                    promises.push(promiseDir(settings.userDir));
+                }
             }
         }
 
@@ -204,7 +206,9 @@ var localfilesystem = {
 
         globalSettingsFile = fspath.join(settings.userDir,".config.json");
 
-        promises.push(promiseDir(libFlowsDir));
+        if (!settings.readOnly) {
+            promises.push(promiseDir(libFlowsDir));
+        }
 
         return when.all(promises);
     },
@@ -227,6 +231,10 @@ var localfilesystem = {
     },
 
     saveFlows: function(flows) {
+        if (settings.readOnly) {
+            return when.resolve();
+        }
+
         if (fs.existsSync(flowsFullPath)) {
             fs.renameSync(flowsFullPath,flowsFileBackup);
         }
@@ -266,6 +274,10 @@ var localfilesystem = {
     },
 
     saveCredentials: function(credentials) {
+        if (settings.readOnly) {
+            return when.resolve();
+        }
+
         if (fs.existsSync(credentialsFile)) {
             fs.renameSync(credentialsFile,credentialsFileBackup);
         }
@@ -297,6 +309,9 @@ var localfilesystem = {
         return when.resolve({});
     },
     saveSettings: function(settings) {
+        if (settings.readOnly) {
+            return when.resolve();
+        }
         return writeFile(globalSettingsFile,JSON.stringify(settings,null,1));
     },
     getSessions: function() {
@@ -317,6 +332,9 @@ var localfilesystem = {
         return when.resolve({});
     },
     saveSessions: function(sessions) {
+        if (settings.readOnly) {
+            return when.resolve();
+        }
         return writeFile(sessionsFile,JSON.stringify(sessions));
     },
 
@@ -364,6 +382,9 @@ var localfilesystem = {
     },
 
     saveLibraryEntry: function(type,path,meta,body) {
+        if (settings.readOnly) {
+            return when.resolve();
+        }
         var fn = fspath.join(libDir, type, path);
         var headers = "";
         for (var i in meta) {
